backend/requirements.txt
asgiref==3.5.0
astroid==2.9.3
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
execnet==1.9.0
filelock==3.4.2
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
pipenv==2022.1.8
platformdirs==2.4.1
pluggy==1.0.0
psycopg==3.0.9
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pylint==2.12.2
pyparsing==3.0.7
pytest==6.2.5
pytest-cache==1.0
pytest-pylint==0.18.0
pytest-subtests==0.6.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.13.0
virtualenv-clone==0.5.7
wrapt==1.13.3
